==> Scripts/subchallenge1/03_probability_HC.ipynb <==
import pandas as pd
from random import randint
import numpy as np

mutation_rate = pd.read_table("./mutation_rate.tsv")
mutation_rate = dict(zip(zip(mutation_rate["state"],mutation_rate["pos"]),mutation_rate["rate"]))
mutation_rate

class TreeNode(object):
    def __init__(self,label=None,seq=None,parent=None,leaves=[]):
        self.label = label
        self.seq = seq
        self.left = None
        self.right = None
        self.parent = parent
        self.leaves = leaves

class hcluster(object):
    def __init__(self,mutation_rate,df):
        self.mutation_rate = mutation_rate
        self.df = df
        self.all_tree_nodes = [TreeNode(label='_'.join([i[1].cell,i[1].state]),seq=i[1].state,\
                                       leaves=[i[0]]) for i in df.iterrows()]
        self.root = TreeNode(leaves=[i for i in range(len(self.all_tree_nodes))])
    
    def initialize_dist_matrix(self,df):
        n = len(df)
        m = len(df["state"][0])
        
        dist_matrix = np.zeros((n,n))
        dist_matrix2 = np.zeros((n,n))
        for i in range(n):
            for j in range(i):
                dist_matrix[i,j] = self.sequence_dist(df["state"][i],df["state"][j])
                dist_matrix2[i,j] = self.sequence_dist2(df["state"][i],df["state"][j])
                
        dist_matrix = dist_matrix + dist_matrix.T
        dist_matrix2 = dist_matrix2 + dist_matrix2.T
        return dist_matrix + self.lambda1 * dist_matrix2
    
    def hamming_dist(self,s1,s2):
        return sum([s1[i]!=s2[i] for i in range(len(s1))])
    
    def hamming_dist2(self,s1,s2):
        return sum([s1[i]!=s2[i] or s1[i]!="1" for i in range(len(s1))])
    
    def sequence_dist(self,s1,s2):
        def match_prob(a,b,i):
            if a == b:
                return math.log(self.mutation_rate[(int(a),i)])
            return 0
        return sum([match_prob(s1[i],s2[i],i) for i in range(len(s1))])
  
    def sequence_dist2(self,s1,s2):
        def mismatch_prob(a,b,i):
            if a == b:
                return 0   
            return -math.log(self.mutation_rate[int(a),i]*self.mutation_rate[int(b),i])
        return sum([mismatch_prob(s1[i],s2[i],i) for i in range(len(s1))])

    
    def merge_node(self,D):
        np.fill_diagonal(D,float("inf"))
        i = np.argmin(D)
        i,j = i//len(D),i%len(D)
        return i,j
    
    def update_dist_matrix(self,clusters):
        def cluster_dist(c1,c2):
            d = 0
            for cc1 in c1:
                for cc2 in c2:
                    d += self.D[cc1,cc2]
            return d/(len(c1)*len(c2))
        
        n = len(clusters)
        D = np.zeros((n,n))
        
        for i in range(n):
            for j in range(i):
                D[i,j] = cluster_dist(clusters[i].leaves,clusters[j].leaves)
                
        D = D + D.T
        return D
    
    def h_clustering(self,lambda1=0.1):
        self.lambda1 = lambda1
        clusters = self.all_tree_nodes
        self.D = self.initialize_dist_matrix(self.df)
        
        D = self.D
        nn = len(self.D)
        
        for c in range(nn-1):
            n = len(D)
            i,j = self.merge_node(D)
            parent = TreeNode()
            parent.left = clusters[i]
            parent.right = clusters[j]
            parent.leaves = clusters[i].leaves + clusters[j].leaves
            clusters = [clusters[idx] for idx in range(len(clusters)) if idx!=i and idx!=j] + [parent]
            D = self.update_dist_matrix(clusters)
        self.root = parent
    
    def write_newick(self):
        def helper(root):
            if not root:
                return

            if not root.left and not root.right:
                return "%s" % (root.label)

            return "(%s,%s)" % (helper(root.left),helper(root.right))

        return helper(self.root) + "root;"

import glob
import numpy as np
import math

def solver(args):
    df = args[0]
    sample = args[1]
    res = hcluster(mutation_rate,df)
    res.h_clustering(lambda1=0)
    outfile = "./%s.nwk" %sample
    f = open(outfile,"w")
    f.write(res.write_newick())
    f.close()

samples = [i.split("/")[-1].split(".")[0] for i in glob.glob("../../Data/Subchallenge1/*.txt")]
for sample in samples:
    if sample:
        df = pd.read_table("../../Data/Subchallenge1/%s.txt"%sample,dtype="str")
        solver([df,sample])

import glob
import numpy as np
import math

def solver(args):
    df = args[0]
    sample = args[1]
    res = hcluster(mutation_rate,df)
    res.h_clustering(lambda1=0)
    outfile = "../../Output/Subchallenge1/HC_hamming/%s.nwk" %sample
    f = open(outfile,"w")
    f.write(res.write_newick())
    f.close()

samples = [i.split("/")[-1].split(".")[0] for i in glob.glob("../../Data/Subchallenge1/*.txt")]
for sample in samples:
    if sample:
        df = pd.read_table("../../Data/Subchallenge1/%s.txt"%sample,dtype="str")
        solver([df,sample])

ID = []
trees = []
samples = [i.split("/")[-1].split(".")[0] for i in glob.glob("../../Data/Subchallenge1/*test*.txt")]
for i in range(30):
    ID.append(i+1)
    sample = "sub1_test_%d"%(i+1)
    df = pd.read_table("../../Data/Subchallenge1/%s.txt"%sample,dtype="str")
    print(sample,len(df))
    res = hcluster(mutation_rate,df)
    res.h_clustering(lambda1=0)
    trees.append(res.write_newick())
    
    outfile = "../../Output/Subchallenge1/HC/%s.nwk" %sample
    f = open(outfile,"w")
    f.write(res.write_newick())
    f.close()

df = pd.DataFrame(data={"dreamID":ID,"nw":trees})
df.to_csv(r'../../Output/Subchallenge1/Jasper_Hu_sub1_submission.txt',index=None, sep='\t')

df
